Geeting data from reducer
