+++ src/components/greetings/Greeting.js
@@ -1,0 +1,7 @@
+const Greeting = () => (
+  <h1>
+    Hello from Greeting Component!
+  </h1>
+);
+
+export default Greeting;
